+++ CHANGELOG.md
@@ -8,6 +8,7 @@ We adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## Unreleased
 
 ### Added
+- Ability to view log stack trace
 - Option to use custom GUI skin (useful to customize scrollbar width)
 
 ### Fixed